*Refactor
- Atomic commits? Never heard of it.
- Extract validation.
- Move control logic into `get_audio.py`
- Add structure folders.
- Get audio methods now in separate classes and modules.
- Move duplicate code into parent class.
- Now user may try to scrape audio if fetch fails.
- Use rich library for better UX.

+++ common/validation.py
@@ -10,9 +10,9 @@ def validate_path(path) -> None:
     if not os.path.exists(path):
         os.makedirs(path)
         print(f'Created directory: "{path}"')
-    if os.path.exists(path) and os.path.isdir(path):
+    if os.path.exists(path) and os.path.isdir(path) and len(os.listdir(path)) != 0:
         x = input(
-            f'[!] Path "{path}" already exists. Script will clear it. Continue? (Y/n): '
+            f'[!] Found files in "{path}". Script will clear it. Continue? (Y/n): '
         ).lower()
         if x != negative_responses:
             shutil.rmtree(path)
